fix: remove duplication of code and make schematics class based
