+++ packages/nest-commander-schematics/package.json
@@ -27,12 +27,13 @@
     "@angular-devkit/core": "11.2.4",
     "@angular-devkit/schematics": "11.2.4",
     "@nestjs/schematics": "^7.3.1",
+    "rxjs": "^6.6.7",
     "typescript": "~4.1.2"
   },
   "devDependencies": {
     "@angular-devkit/schematics-cli": "^0.1102.8",
-    "@types/node": "^14.14.37",
     "@types/jest": "^26.0.22",
+    "@types/node": "^14.14.37",
     "jest": "^26.6.3",
     "ts-jest": "^26.5.4"
   }
