+++ packages/nest-commander-schematics/src/command/command-options.interface.ts
@@ -1,14 +1,7 @@
-import { Path } from '@angular-devkit/core';
+import { CommonOptions } from '../common/common-options.interface';
 
-export interface CommandOptions {
-  name: string;
+export interface CommandOptions extends CommonOptions {
   question: string;
   default: boolean;
-  spec: boolean;
-  flat: boolean;
-  sourceRoot: string;
-  path?: string | Path;
-  type?: string;
-  metadata?: string;
-  module?: Path | null;
+  isDefault?: boolean;
 }
